# contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "momiao owned APIs",
    "version": "0.4.0",
    "description": "Owned process liveness and authenticated local wallet APIs. Native /api/, /v1/ and /pg contracts remain external. Wallet configuration is optional; no payment, grant, transfer or ledger mutation endpoint is exposed.",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Default local development listener"
    }
  ],
  "security": [],
  "paths": {
...
    "/platform/v1/wallet": {
      "get": {
        "operationId": "getWallet",
        "security": [
          {
            "nativeBearer": [],
            "nativeUser": []
          }
        ],
        "parameters": [
          {
            "name": "X-Auth-Session",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verified user local-wallet result. No total-asset completeness assertion.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/WalletProjection"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or duplicate/unknown fields",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or rejected identity",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletError"
                }
              }
            }
          },
          "403": {
            "description": "Native permission or Origin rejected; do not clear valid session",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletError"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletError"
                }
              }
            }
          },
          "502": {
            "description": "Native identity service unavailable or malformed",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletError"
                }
              }
            }
          },
          "503": {
            "description": "Wallet not enabled, unavailable or incomplete",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletError"
                }
              }
            }
          }
        },
        "description": "No query parameters; no writes. Zero rows means uninitialized. One missing wallet row is an error, not a zero balance."
      }
    },
...
    "/platform/v1/master-profile": {
      "get": {
        "operationId": "getMasterProfile",
        "security": [
          {
            "nativeBearer": [],
            "nativeUser": []
          }
        ],
        "parameters": [
          {
            "name": "X-Auth-Session",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Own Master profile; all IDs/versions are exact strings.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/MasterProfile"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or duplicate/unknown fields",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or rejected identity",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterError"
                }
              }
            }
          },
          "403": {
            "description": "Native authorization, exact Origin, or reserved nickname rejected; does not imply session logout",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterError"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterError"
                }
              }
            }
          },
          "502": {
            "description": "Native identity service unavailable or malformed",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterError"
                }
              }
            }
          },
          "503": {
            "description": "Profile store disabled or unavailable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterError"
                }
              }
            }
          }
        },
        "description": "No query parameters. Every request verifies native identity. Only own display identity; no native credentials, wallet writes or grants."
      },
      "patch": {
        "operationId": "updateMasterProfile",
        "security": [
          {
            "nativeBearer": [],
            "nativeUser": []
          }
        ],
        "parameters": [
          {
            "name": "X-Auth-Session",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Exact configured HTTPS origin."
          }
        ],
        "responses": {
          "200": {
            "description": "Own Master profile; all IDs/versions are exact strings.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/MasterProfile"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or duplicate/unknown fields",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or rejected identity",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterError"
                }
              }
            }
          },
          "403": {
            "description": "Native authorization, exact Origin, or reserved nickname rejected; does not imply session logout",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterError"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterError"
                }
              }
            }
          },
          "502": {
            "description": "Native identity service unavailable or malformed",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterError"
                }
              }
            }
          },
          "503": {
            "description": "Profile store disabled or unavailable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterError"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type must be application/json",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterError"
                }
              }
            }
          },
          "409": {
            "description": "NICKNAME_TAKEN, STALE_RESOURCE_VERSION or RENAME_COOLDOWN",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "expected_version"
                ],
                "properties": {
                  "expected_version": {
                    "type": "string",
                    "pattern": "^[1-9][0-9]*$"
                  },
                  "display_name": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Server validates normalized 1..24 Unicode grapheme clusters, allowed characters and reserved names."
                  },
                  "avatar_id": {
                    "type": "string",
                    "const": "system-default"
                  }
                },
                "anyOf": [
                  {
                    "required": [
                      "display_name"
                    ]
                  },
                  {
                    "required": [
                      "avatar_id"
                    ]
                  }
                ]
              }
            }
          }
        },
        "description": "No query parameters. Every request verifies native identity. Only own display identity; no native credentials, wallet writes or grants. Strict JSON; duplicate/unknown/null fields rejected. Unknown write outcome requires GET reconciliation, not automatic replay."
      }
    },
...
    "/platform/v1/transactions": {
      "get": {
        "operationId": "listAssetTransactions",
        "security": [
          {
            "nativeBearer": [],
            "nativeUser": []
          }
        ],
        "description": "Owner-only local wallet operation. No native quota movement. See wallet-actions-api.md for business and retry semantics.",
        "parameters": [
          {
            "name": "X-Auth-Session",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            }
          },
          {
            "name": "after_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f-]{36}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Confirmed receipt or read-only state; Cache-Control: no-store.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "const": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/AssetTransactionPage"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Structured error; no-store; no raw internal details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EconomyError"
                }
              }
            }
          },
          "401": {
            "description": "Structured error; no-store; no raw internal details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EconomyError"
                }
              }
            }
          },
          "403": {
            "description": "Structured error; no-store; no raw internal details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EconomyError"
                }
              }
            }
          },
          "405": {
            "description": "Structured error; no-store; no raw internal details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EconomyError"
                }
              }
            }
          },
          "409": {
            "description": "Structured error; no-store; no raw internal details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EconomyError"
                }
              }
            }
          },
          "415": {
            "description": "Structured error; no-store; no raw internal details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EconomyError"
                }
              }
            }
          },
          "502": {
            "description": "Structured error; no-store; no raw internal details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EconomyError"
                }
              }
            }
          },
          "503": {
            "description": "Structured error; no-store; no raw internal details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EconomyError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "nativeBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "nativeUser": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User"
      }
    },
    "schemas": {
      "WalletProjection": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "initialized",
          "user_id",
          "wallets",
          "scope",
          "total_assets"
        ],
        "properties": {
          "initialized": {
            "type": "boolean"
          },
          "user_id": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "wallets": {
            "type": "array",
            "maxItems": 2,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "asset",
                "balance_units",
                "amount",
                "ledger_seq",
                "version"
              ],
              "properties": {
                "asset": {
                  "type": "string",
                  "enum": [
                    "RESERVE_API_CREDIT",
                    "AVAILABLE_CHIPS"
                  ]
                },
                "balance_units": {
                  "type": "string",
                  "pattern": "^(0|[1-9][0-9]*)$"
                },
                "amount": {
                  "type": "string",
                  "pattern": "^-?[0-9]+(\\.[0-9]{1,6})?$"
                },
                "ledger_seq": {
                  "type": "string",
                  "pattern": "^(0|[1-9][0-9]*)$"
                },
                "version": {
                  "type": "string",
                  "pattern": "^(0|[1-9][0-9]*)$"
                }
              }
            }
          },
          "scope": {
            "type": "string",
            "const": "LOCAL_WALLETS_ONLY"
          },
          "total_assets": {
            "type": "null"
          }
        }
      },
...
      "WalletError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "success",
          "code",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "MasterError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "success",
          "code",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "EconomyError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "success",
          "code",
          "message"
        ],
        "properties": {
          "success": {
            "const": false
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}